Update and rename README.md to CODE
This is a project presented in SQL, that represents relationships and formulations in the COVID dataset—using the website ourworldindata.org, to find the dataset. The dataset consists of attributes like Location, date, total deaths, total cases, and so on. After getting the data, using excel data cleaning was performed. Then using SQL, explore the dataset in order to find valuable advances.